news_pipeline/requirements.txt
redis
python-dateutil
numpy
scipy
sklearn
tap-news-utils
